requirements.txt
boto3==1.24.34
botocore==1.27.34
certifi==2022.6.15
charset-normalizer==2.1.0
cronitor==4.5.0
humanize==4.2.3
idna==3.3
jmespath==1.0.1
numpy==1.23.1
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.10
PIconnect
